05/19/26

+++ pages/tools/remove_replaced_files.php
@@ -66,11 +66,14 @@ foreach ($resource_alt_files as $alt_file) {
         if ($dry_run) {
             echo "Resource {$alt_file['resource']} - Found alternative file $alt_file_location" . PHP_EOL;
         } else {
-            unlink($alt_file_location);
-            echo "Resource {$alt_file['resource']} - Deleted alternative file $alt_file_location" . PHP_EOL;
-            ps_query("UPDATE resource_log SET previous_file_alt_ref = NULL WHERE ref = ?;" , array('i', $alt_file['resource_log_ref'])); # Don't show revert option in resource log.
-            ps_query("DELETE FROM resource_alt_files WHERE ref = ?;", array('i', $alt_file['previous_file_alt_ref']));
-            $update_disk_usage = true;
+            if (try_unlink($alt_file_location) === true) {
+                echo "Resource {$alt_file['resource']} - Deleted alternative file $alt_file_location" . PHP_EOL;
+                ps_query("UPDATE resource_log SET previous_file_alt_ref = NULL WHERE ref = ?;" , array('i', $alt_file['resource_log_ref'])); # Don't show revert option in resource log.
+                ps_query("DELETE FROM resource_alt_files WHERE ref = ?;", array('i', $alt_file['previous_file_alt_ref']));
+                $update_disk_usage = true;
+            } else {
+                echo "ERROR - Resource {$alt_file['resource']} - An error occurred when deleting alternative file $alt_file_location" . PHP_EOL;
+            }
         }
     }
 
@@ -79,9 +82,12 @@ foreach ($resource_alt_files as $alt_file) {
         if ($dry_run) {
             echo "Resource {$alt_file['resource']} - Found alternative file thumb $alt_file_thm_location" . PHP_EOL;
         } else {
-            unlink($alt_file_thm_location);
-            echo "Resource {$alt_file['resource']} - Deleted alternative file thumb $alt_file_thm_location" . PHP_EOL;
-            $update_disk_usage = true;
+            if (try_unlink($alt_file_thm_location) === true) {
+                echo "Resource {$alt_file['resource']} - Deleted alternative file thumb $alt_file_thm_location" . PHP_EOL;
+                $update_disk_usage = true;
+            } else {
+                echo "ERROR - Resource {$alt_file['resource']} - An error occurred when deleting alternative file thumb $alt_file_thm_location" . PHP_EOL;
+            }
         }
     }
